Added depth trails and shortest path

+++ scripts/algorithms/tpbp_alt/tpbp_alt1.py
@@ -29,9 +29,8 @@ import numpy as np
 
 
 def add_vertex_trail(graph, path, len_path, vertex, dest, len_max):
-    ''''''
     cur = path[-1]
-    len_rem = nx.dijkstra_path_length(graph, cur, dest, weight = 'length')
+    len_rem = nx.dijkstra_path_length(graph, cur, vertex, weight = 'length')
     if (len_rem + len_path) > len_max:
         return False
     if not cur in path[:-1]:
@@ -41,11 +40,12 @@ def add_vertex_trail(graph, path, len_path, vertex, dest, len_max):
             return False
     return True
 
-def compute_valid_trails(graph, source, dest, len_max, folder):
-    with open(folder + '/valid_trails_{}_{}_{}.in'.format(str(source), str(dest),str(int(len_max))), 'w') as f:
+def compute_valid_trails(graph, source, dest, len_max, depth, folder):
+    #if os.path.exists(folder + '/depth_trails_{}_{}_{}.in'.format(str(source), str(dest),str(int(len_max)/10+1))):
+    #    return
+    with open(folder + '/depth_trails_{}_{}_{}.in'.format(str(source), str(dest), str(depth)), 'w') as f:
         with open(folder + '/vp_temp_{}.in'.format(0), 'w') as f1:
             f1.write(str(source) + ' ' + str(0) + '\n')
-
         count = 1  #no of walks in vp temp
         steps = 0   # iterations on vp temp
         while count != 0:
@@ -65,28 +65,39 @@ def compute_valid_trails(graph, source, dest, len_max, folder):
                             if add_vertex_trail(graph, path, len_path, v, dest, len_max * 10.):
                                 temp = ' '.join(line1[:-1])
                                 temp = temp + ' ' + str(v)
-                                if v == dest:
+                                if v==dest:
                                     f.write(temp + '\n')
+                                    continue
                                 else:
                                     count += 1
                                     temp += ' ' + str(graph[path[-1]][v]['length'] + len_path)
                                     f1.write(temp + '\n')
             steps += 1
+            if steps >= depth:
+                with open(folder + '/vp_temp_{}.in'.format(steps), 'r') as f0:
+                    for line in f0:
+                        line1 = line.split('\n')
+                        line_temp = line1[0]
+                        line1 = line_temp.split(' ')
+                        path = ' '.join(line1[:-1])
+                        f.write(path+'\n')
+                        
+                break
             os.remove(folder + '/vp_temp_{}.in'.format(steps-1))
-
+    os.remove(folder + '/vp_temp_{}.in'.format(steps))
     #for i in range(steps + 1):
     #    os.remove(folder + '/vp_temp_{}.in'.format(i))
 
-def all_valid_trails(graph, node_set, len_max, folder):
+def all_valid_trails(graph, node_set, len_max, depth, folder):
     #suggestion using j in range(i+1,len(node_set)) might help reducing time complexity to n(n-1)/2, current time complexity is n^2
     for i in range(len(node_set)):
         for j in range(len(node_set)):
-            compute_valid_trails(graph, node_set[i], node_set[j], len_max[i], folder)
+            compute_valid_trails(graph, node_set[i], node_set[j], len_max[i], depth, folder)
 
 
 class TPBP:
 
-    def __init__(self, graph, priority_nodes, time_periods, coefficients, path_to_folder):
+    def __init__(self, graph, priority_nodes, time_periods, coefficients, depth, path_to_folder):
         '''Initializes class TPBP with required ros paramenters  i.e time periods,dummy nodes reshuffle time,etc '''
         self.ready = False
         rospy.Service('algo_ready', AlgoReady, self.callback_ready)
@@ -94,6 +105,7 @@ class TPBP:
         self.priority_nodes = priority_nodes     
         self.time_periods = time_periods
         self.coefficients = coefficients
+        self.depth = depth
         self.offline_folder = path_to_folder   #offline folder path to store valid walks
         for node in self.graph.nodes():
             self.graph.nodes[node]['idleness'] = 0.     #adding a idleness parameter to nodes
@@ -130,7 +142,7 @@ class TPBP:
         for _ in range(s):
             ''' assigning time periods to non priority nodes'''
             temp.append(self.time_periods[0])
-        all_valid_trails(self.graph, self.priority_nodes, temp, self.offline_folder)
+        all_valid_trails(self.graph, self.priority_nodes, temp, self.depth, self.offline_folder)
         time.sleep(1.)  #halts the exceution of code for 1 sec
         self.ready = True
 
@@ -146,6 +158,7 @@ class TPBP:
                     temp2 += max(self.graph.nodes[i]['idleness'] - self.time_periods[j], 0)
 
         temp3 = 0.
+        '''
         for i in range(len(self.priority_nodes)):
             if not self.assigned[i] and not self.priority_nodes[i] in nodes:
                 dist = np.inf
@@ -153,10 +166,11 @@ class TPBP:
                     temp = nx.dijkstra_path_length(self.graph, j, self.priority_nodes[i], 'length')
                     if temp < dist:
                         dist = temp
-                temp3 += dist
+                temp3 += dist'''
         temp4 = 0
+        '''
         for i in range(len(walk) - 1):
-            temp4 += self.graph[walk[i]][walk[i + 1]]['length']
+            temp4 += self.graph[walk[i]][walk[i + 1]]['length']'''
         
         return np.dot([temp1, temp2, temp3, temp4], coefficients)
 
@@ -208,13 +222,17 @@ class TPBP:
         for j in range(len(self.priority_nodes)):
             if not self.assigned[j]:
                 '''211 kuch aur ayega'''
-                valid_trails = '/valid_trails_{}_{}_{}.in'.format(node, self.priority_nodes[j], str(int(self.time_periods[j])))
+                valid_trails = '/depth_trails_{}_{}_{}.in'.format(node, self.priority_nodes[j], str(self.depth))
                 with open(self.offline_folder + valid_trails, 'r') as f:
                     count = 0
                     for line in f:
                         count += 1
                         line1 = line.split('\n')
                         line2 = line1[0].split(' ')
+                        line3 = nx.dijkstra_path(self.graph, line2[-1], self.priority_nodes[j], weight='length')
+                        if line2[-1] not in self.priority_nodes:
+                            line2.extend(line3[1:])
+                            print('line2 after extension',line2)
                         r = self.tpbp_reward(line2)
                         if r > best_reward:
                             best_reward = r
@@ -258,7 +276,7 @@ class TPBP:
 
     def callback_ready(self, req):
         algo_name = req.algo
-        if algo_name == 'tpbp_final' and self.ready:
+        if algo_name == 'tpbp_alt1' and self.ready:
             return AlgoReadyResponse(True)
         else:
             return AlgoReadyResponse(False)
@@ -269,16 +287,17 @@ if __name__ == '__main__':
     done = False
     graph_name = rospy.get_param('/graph')
     g = nx.read_graphml(dirname + '/graph_ml/' + graph_name + '.graphml')
-
+    algo_name=rospy.get_param('/algo_name')
     priority_nodes = rospy.get_param('/priority_nodes').split(' ')
     time_periods = list(map(float, rospy.get_param('/time_periods').split(' ')))
     coefficients = list(map(float, rospy.get_param('/coefficients').split(' ')))
+    depth = rospy.get_param('/depth')
     #folder = rospy.get_param('/random_string')
-    folder = 'tpbp'
+    folder = algo_name + graph_name
     #num_dummy_nodes = rospy.get_param('/num_dummy_nodes')
     #reshuffle_time = rospy.get_param('/reshuffle_time')
     path_to_folder = dirname + '/outputs/' + folder
-    s = TPBP(g, priority_nodes, time_periods, coefficients, path_to_folder)
+    s = TPBP(g, priority_nodes, time_periods, coefficients, depth, path_to_folder)
 
     rospy.Subscriber('at_node', AtNode, s.callback_idle)
     rospy.Service('bot_next_task', NextTaskBot, s.callback_next_task)
